+++ CHANGELOG.md
@@ -40,6 +40,7 @@ All notable changes to this project will be documented in this file.
 - Added Health and Mana Regen flesh heaps
 *
 - Fixed the Spawn Camps OP versions, reduced manacost of 100 for all of them.
+- Maybe fixed gyro missle 'no charges' bug, may require respawning to fix